feat: alteracoes de sintaxe

+++ Store/src/main/java/com/abila/Store/domain/DTO/ItemVendaRequest.java
@@ -4,9 +4,11 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 
 public record ItemVendaRequest(
         @NotBlank String nome,
-        @NotNull @Positive Float precoUnitario,
+        @NotNull @Positive BigDecimal precoUnitario,
         @NotNull @Positive Integer quantidade
 ) { }
