+++ Store/src/main/java/com/abila/Store/domain/DTO/ItemVendaResponse.java
@@ -3,10 +3,12 @@ package com.abila.Store.domain.DTO;
 import com.abila.Store.domain.ItemVendas;
 import jakarta.persistence.Column;
 
+import java.math.BigDecimal;
+
 public record ItemVendaResponse(
          Integer id,
          String nome,
-         Float precoUnitario,
+         BigDecimal precoUnitario,
          Integer quantidade
 ) {
     public ItemVendaResponse(ItemVendas item){
